Refactor and enhance camera settings and frame processing
Refactored frame processing to use stack-allocated DetectorEnvironment and updated mutex handling in BaseQueue.h. Enhanced camera settings with a new popup modal, including functions to display and hide it. Improved type safety in JsonWriter.cpp and added GPU usage functions in system_usage.cpp. Various null checks and logging improvements across multiple files.

+++ cs_vision_common/JsonWriter.cpp
@@ -57,8 +57,8 @@ int JsonWriter::set_json_value(const char* path, json_value value)
 
 int JsonWriter::get_json_element_index(string& name)
 {
-	int i = name.find_last_of('[');
-	int j = name.find_last_of(']');
+	size_t i = name.find_last_of('[');
+	size_t j = name.find_last_of(']');
 	if (i == string::npos || j == string::npos || i >= j)
 		return -1;
